requirements.txt
apiclient==1.0.4
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.2
Flask==1.1.2
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.17.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
wincertstore==0.2
PyPDF2
python-docx
gunicorn
nltk==3.6.2
numpy==2.2.4
Pillow
opencv-python==4.11.0.86
imageio==2.37.0 
lazy-loader==0.4 
networkx==3.4.2 
scikit-image==0.25.2 
scipy==1.15.2 
tifffile==2025.3.30